requirements.txt
aiodns==3.0.0; python_version >= "3.6"
aiohttp==3.8.1; python_version >= "3.6"
aioredis==2.0.1; python_version >= "3.7" and python_version < "4.0"
aiosignal==1.2.0; python_version >= "3.6"
amogus==0.1
anyio==3.5.0; python_version >= "3.6" and python_full_version >= "3.6.2"
asgiref==3.4.1; python_version >= "3.6"
async-timeout==4.0.2; python_version >= "3.7" and python_version < "4.0"
attrs==21.4.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
brotli==1.0.9; python_version >= "3.6"
cchardet==2.1.7; python_version >= "3.6"
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
cffi==1.15.0; python_version >= "3.6"
charset-normalizer==2.0.10; python_full_version >= "3.6.0" and python_version >= "3.6"
click==8.0.3; python_version >= "3.6"
colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.5.0" and platform_system == "Windows"
fastapi-limiter @ git+https://github.com/long2ice/fastapi-limiter.git@master ; python_version >= "3.7" and python_version < "4.0"
fastapi==0.70.1; python_full_version >= "3.6.1"
frozenlist==1.3.0; python_version >= "3.7"
h11==0.13.0; python_version >= "3.6"
idna==3.3; python_full_version >= "3.6.2" and python_version >= "3.6"
loguru==0.5.3; python_version >= "3.5"
multidict==5.2.0; python_version >= "3.6"
numpy==1.22.1; python_version >= "3.8"
opencv-python-headless==4.5.4.60; python_version >= "3.6"
pillow==8.2.0; python_version >= "3.6"
pycares==4.1.2; python_version >= "3.6"
pycparser==2.21; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.4.0"
pydantic==1.9.0; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.1"
pytesseract==0.3.8; python_version >= "3.6"
pyyaml==6.0; python_version >= "3.6"
requests==2.27.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
sniffio==1.2.0; python_version >= "3.6" and python_full_version >= "3.6.2"
starlette==0.16.0; python_version >= "3.7" and python_full_version >= "3.6.1" and python_version < "4.0"
typing-extensions==4.0.1; python_version >= "3.7" and python_full_version >= "3.6.1" and python_version < "4.0"
urllib3==1.26.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
uvicorn==0.16.0
win32-setctime==1.0.4; sys_platform == "win32" and python_version >= "3.5"
yarl==1.7.2; python_version >= "3.6"
